rework logstore interfaces, break off from kevalue, bugfixes
