+++ ee/agent/types/keyvalue_store.go
@@ -24,11 +24,6 @@ type Deleter interface {
 	DeleteAll() error
 }
 
-// RowDeleter is an interface for deleting rows by rowid in a sql store
-type RowDeleter interface {
-	DeleteRows(rowids ...any) error
-}
-
 // Iterator is an interface for iterating data in a key/value store.
 type Iterator interface {
 	// ForEach executes a function for each key/value pair in a store.
@@ -36,21 +31,6 @@ type Iterator interface {
 	// the error is returned to the caller. The provided function must not modify
 	// the store; this will result in undefined behavior.
 	ForEach(fn func(k, v []byte) error) error
-}
-
-// TimestampedIterator is a read-only interface for iterating timestamped data.
-type TimestampedIterator interface {
-	// ForEach executes a function for each timestamp/value pair in a store.
-	// If the provided function returns an error then the iteration is stopped and
-	// the error is returned to the caller. The provided function must not modify
-	// the store; this will result in undefined behavior.
-	ForEach(fn func(rowid, timestamp int64, v []byte) error) error
-}
-
-// TimestampedAppender is an interface for supporting the addition of timestamped values to a store
-type TimestampedAppender interface {
-	// AppendValue takes the timestamp, and marshalled value for insertion as a new row
-	AppendValue(timestamp int64, value []byte) error
 }
 
 // Updater is an interface for bulk replacing data in a key/value store.
@@ -122,20 +102,6 @@ type GetterSetterDeleterIteratorUpdaterCounterAppender interface {
 	Counter
 	Appender
 }
-
-// TimestampedIteratorDeleterAppenderCounterCloser is an interface to support the storage and retrieval of
-// sets of timestamped values. This can be used where a strict key/value interface may not suffice,
-// e.g. for writing logs or historical records to sqlite
-type TimestampedIteratorDeleterAppenderCounterCloser interface {
-	TimestampedIterator
-	TimestampedAppender
-	Counter
-	RowDeleter
-	Closer
-}
-
-// LogStore is a convenient alias for a store that supports all methods required to manipulate sqlite logs
-type LogStore = TimestampedIteratorDeleterAppenderCounterCloser
 
 // Convenient alias for a key value store that supports all methods
 type KVStore = GetterSetterDeleterIteratorUpdaterCounterAppender
